box: move styles+icons to frontend for bundle size
