+++ src/edtr-io/plugins/box/renderer.tsx
@@ -1,41 +1,11 @@
-import { faLightbulb, faQuoteRight } from '@edtr-io/ui'
-import {
-  faScroll,
-  faHandPointRight,
-  faMapSigns,
-  faThumbtack,
-  faExclamationTriangle,
-} from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import clsx from 'clsx'
 
 import { BoxProps } from '.'
+import { boxTypeStyle, defaultStyle } from '@/components/content/box'
 import { useInstanceData } from '@/contexts/instance-context'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
 import { hasOwnPropertyTs } from '@/helper/has-own-property-ts'
-
-export const boxTypeStyle = {
-  blank: {},
-  example: {},
-  quote: { icon: faQuoteRight },
-  approach: { icon: faMapSigns },
-  remember: { icon: faScroll },
-  attention: {
-    icon: faExclamationTriangle,
-    borderColorClass: 'border-red-100',
-    colorClass: 'text-orange',
-  },
-  note: { icon: faHandPointRight },
-  definition: { icon: faThumbtack },
-  theorem: { icon: faLightbulb },
-  proof: {},
-}
-
-export const defaultStyle = {
-  icon: undefined,
-  borderColorClass: 'border-brand-300',
-  colorClass: 'text-brand',
-}
 
 const types = Object.keys(boxTypeStyle)
 export type BoxType = keyof typeof boxTypeStyle
